Move Scene and getScenes to utils.ts
